update: Add ViewModels to sync also load locally from the database Projects also ArmsTeam Entities.
update: Remove unnecessary local data from Local.kts.

feat: Add ViewModelFacory pattern.

+++ app/src/main/java/com/example/armsapp/data/local/entities/ArmsTeamEntity.kt
@@ -14,11 +14,10 @@ data class ArmsTeamEntity(
     val image: String
 )
 
-//TODO: Arrumar as propriedade para String
 fun ArmsTeamEntity.toArmsTeam() = ArmsTeam(
     id = id,
-    name = name.toInt(),
-    jobPosition = jobPosition.toInt(),
+    name = name,
+    jobPosition = jobPosition,
     instagramLabel = instagramLabel,
     instagramUrl = instagramUrl,
     imageUrl = image
